:hammer: chore: refactor hooks/ folder
