+++ packages/client/src/components/dispatch/ActiveDeputies.tsx
@@ -5,7 +5,7 @@ import { ManageUnitModal } from "./modals/ManageUnit";
 import { useModal } from "context/ModalContext";
 import { ModalIds } from "types/ModalIds";
 import { ActiveDeputy } from "state/emsFdState";
-import { useActiveDeputies } from "hooks/useActiveDeputies";
+import { useActiveDeputies } from "hooks/realtime/useActiveDeputies";
 import { useRouter } from "next/router";
 import { formatUnitDivisions, makeUnitName } from "lib/utils";
 import { useGenerateCallsign } from "hooks/useGenerateCallsign";
