+++ packages/client/src/hooks/shared/useTime.ts
@@ -1,5 +1,5 @@
-import format from "date-fns/format";
 import * as React from "react";
+import format from "date-fns/format";
 
 export function useTime() {
   const ref = React.useRef<HTMLSpanElement>(null);
